move Process Cancellation and Piping code to ProcessWrapper
