--- src/CliInvoke/Helpers/Processes/Cancellation/WindowsGracefulCancellation.cs
@@ -1,113 +1,0 @@
-/*
-    CliInvoke
-    Copyright (C) 2024-2026  Alastair Lundy
-
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at http://mozilla.org/MPL/2.0/.
-   */
-
-using System.Runtime.InteropServices;
-// ReSharper disable PartialTypeWithSinglePart
-
-namespace CliInvoke.Helpers.Processes.Cancellation;
-
-internal static partial class WindowsGracefulCancellation
-{
-    private const uint CtrlCSignalEvent = 0;
-
-    extension(ProcessWrapper process)
-    {
-        /// <summary>
-        /// </summary>
-        /// <param name="timeoutThreshold"></param>
-        /// <param name="exitConfiguration"></param>
-        /// <param name="cancellationToken"></param>
-        /// <exception cref="PlatformNotSupportedException"></exception>
-        [SupportedOSPlatform("windows")]
-        internal async Task<bool> CancelWithInterruptOnWindows(TimeSpan timeoutThreshold,
-            ProcessExitConfiguration exitConfiguration, CancellationToken cancellationToken)
-        {
-            if (!OperatingSystem.IsWindows())
-                throw new PlatformNotSupportedException();
-
-            // Use semaphore to prevent simultaneous cancellation attempts
-            if (!await process._cancellationSemaphore.WaitAsync(0, cancellationToken))
-            {
-                // Another cancellation is already in progress, wait for it to complete
-                await process.WaitForExitAsync(cancellationToken);
-                return process.HasExited;
-            }
-
-            try
-            {
-                DateTime expectedExitTime =
-                    CancellationHelper.CalculateExpectedExitTime(exitConfiguration);
-
-                // Use a local variable to store the cancellation reason to avoid race conditions
-                CancellationReason cancellationReason = CancellationReason.NotKnown;
-
-                // Register the callback to update the cancellation reason
-                cancellationToken.Register(() =>
-                {
-                    cancellationReason =
-                        CancellationHelper.GetCancellationReason(expectedExitTime,
-                            cancellationToken);
-                });
-
-                bool ctrlCSignalSuccess = false;
-
-                try
-                {
-                    await Task.Delay(timeoutThreshold, cancellationToken);
-
-                    if (process.HasExited)
-                        return true;
-
-                    // Allocate a Console to the Process so that it has one it can use.
-                    bool successfulAttachment = AllocConsoleWin();
-                    // Attach the allocated console to the process.
-                    successfulAttachment =
-                        successfulAttachment && AttachConsoleWin((uint)process.Id);
-
-                    if (!successfulAttachment)
-                        return false;
-
-                    ctrlCSignalSuccess = SendCtrlCToConsoleWin(CtrlCSignalEvent, 0);
-                }
-                catch (Exception exception)
-                {
-                    // Recalculate expected exit time in exception handler to avoid using stale values
-                    DateTime currentExpectedExitTime =
-                        CancellationHelper.CalculateExpectedExitTime(exitConfiguration);
-                    CancellationHelper.HandleCancellationExceptions(currentExpectedExitTime,
-                        cancellationReason,
-                        exitConfiguration, exception);
-                }
-                finally
-                {
-                    if (!process.HasExited)
-                        process.ForcefulExit();
-                }
-
-                return ctrlCSignalSuccess;
-            }
-            finally
-            {
-                process._cancellationSemaphore.Release();
-            }
-        }
-    }
-    
-    [LibraryImport("Kernel32.dll", EntryPoint = "GenerateConsoleCtrlEvent", SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static partial bool SendCtrlCToConsoleWin(uint ctrlEvent, uint processGroupEventId);
-
-    [LibraryImport("Kernel32.dll", EntryPoint = "AllocConsole", SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static partial bool AllocConsoleWin();
-
-    [LibraryImport("Kernel32.dll", EntryPoint = "AttachConsole", SetLastError = true)]
-    [return: MarshalAs(UnmanagedType.Bool)]
-    private static partial bool AttachConsoleWin(uint processId);
-}
