--- src/CliInvoke/Helpers/Processes/ProcessPipingHelper.cs
@@ -1,84 +1,0 @@
-﻿/*
-    CliInvoke
-    Copyright (C) 2024-2026  Alastair Lundy
-
-    This Source Code Form is subject to the terms of the Mozilla Public
-    License, v. 2.0. If a copy of the MPL was not distributed with this
-    file, You can obtain one at http://mozilla.org/MPL/2.0/.
-   */
-
-namespace CliInvoke.Helpers.Processes;
-
-/// <summary>
-///     An implementation of IProcessPipeHandler. Pipes Process Standard Input, Output, and Error as
-///     required.
-/// </summary>
-internal static class ProcessPipingHelper
-{
-    /// <param name="destination">The process to which the standard input will be piped.</param>
-    extension(Process destination)
-    {
-        /// <summary>
-        ///     Asynchronously pipes the standard input from a source stream to a specified process.
-        /// </summary>
-        /// <param name="source">The stream from which to read the standard input data.</param>
-        /// <param name="cancellationToken"></param>
-        /// <returns>A task that represents the asynchronous operation containing the destination process.</returns>
-        [UnsupportedOSPlatform("ios")]
-        [UnsupportedOSPlatform("tvos")]
-        [UnsupportedOSPlatform("browser")]
-        internal async Task<bool> PipeStandardInputAsync(Stream source,
-            CancellationToken cancellationToken)
-        {
-            if (destination.StartInfo.RedirectStandardInput)
-            {
-                await destination.StandardInput.FlushAsync(cancellationToken);
-                destination.StandardInput.BaseStream.Position = 0;
-                await source.CopyToAsync(destination.StandardInput.BaseStream, cancellationToken);
-
-                return source.Equals(destination.StandardInput.BaseStream);
-            }
-
-            return false;
-        }
-    }
-    extension(Process source){
-        /// <summary>
-        ///     Asynchronously retrieves the standard output stream from a specified process.
-        /// </summary>
-        /// <param name="cancellationToken"></param>
-        /// <returns>A task that represents the asynchronous operation, containing the standard output stream.</returns>
-        [UnsupportedOSPlatform("ios")]
-        [UnsupportedOSPlatform("tvos")]
-        [UnsupportedOSPlatform("browser")]
-        internal async Task<Stream> PipeStandardOutputAsync(CancellationToken cancellationToken)
-        {
-            Stream destination = new MemoryStream();
-
-            if (source.StartInfo.RedirectStandardOutput)
-                if (source.StandardOutput != StreamReader.Null)
-                    await source.StandardOutput.BaseStream.CopyToAsync(destination, cancellationToken);
-
-            return destination;
-        }
-
-        /// <summary>
-        ///     Asynchronously retrieves the standard error stream from a specified process.
-        /// </summary>
-        /// <param name="cancellationToken"></param>
-        /// <returns>A task that represents the asynchronous operation, containing the standard error stream.</returns>
-        [UnsupportedOSPlatform("ios")]
-        [UnsupportedOSPlatform("tvos")]
-        [UnsupportedOSPlatform("browser")]
-        internal async Task<Stream> PipeStandardErrorAsync(CancellationToken cancellationToken)
-        {
-            Stream destination = new MemoryStream();
-
-            if (source.StartInfo.RedirectStandardError)
-                if (source.StandardError != StreamReader.Null)
-                    await source.StandardError.BaseStream.CopyToAsync(destination, cancellationToken);
-
-            return destination;
-        }
-    }
-}
