+++ src/CliInvoke/Processes/ExternalProcess.cs
@@ -10,6 +10,7 @@
 using CliInvoke.Core.Processes;
 using CliInvoke.Helpers;
 using CliInvoke.Helpers.Processes;
+using CliInvoke.Processes.Internal;
 
 namespace CliInvoke.Processes;
